img_processing: organize directories
