+++ img_processing/include/detection_object_struct.h
