+++ img_processing/include/dynamic_tracker.h
