+++ img_processing/include/manager.h
