+++ img_processing/include/object_type_enum.h
