+++ img_processing/include/tracker_object_struct.h
