+++ img_processing/src/detector.cpp
@@ -1,5 +1,4 @@
-#include "detector.h"
-
+#include "../include/detector.h"
 using namespace std;
 using namespace cv;
 using namespace dnn;
