+++ img_processing/src/dynamic_tracker.cpp
@@ -1,7 +1,7 @@
 #include <opencv2/tracking.hpp>
 #include <opencv2/core/ocl.hpp>
 #include <opencv2/opencv.hpp>
-#include "dynamic_tracker.h"
+#include "../include/dynamic_tracker.h"
 using namespace std;
 using namespace cv;
 using namespace chrono;
