+++ img_processing/src/manager.cpp
@@ -1,4 +1,4 @@
-#include "manager.h"
+#include "../include/manager.h"
 using namespace std;
 using namespace cv;
