+++ img_processing/tests/images/three_cars_image.png
