+++ img_processing/tests/images/track_2_cars_first_frame.jpg
