+++ img_processing/tests/images/track_2_cars_second_frame.jpg
