+++ img_processing/tests/images/two_cars_image.png
